# frontend/lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Intuitiveness API",
    "summary": "REST surface over the descent/ascent abstraction-level engine.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/sessions": {
      "get": {
        "tags": [
          "sessions"
        ],
        "summary": "List Sessions",
        "description": "Index of persisted sessions (id + title).",
        "operationId": "list_sessions_sessions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SessionSummary"
                  },
                  "type": "array",
                  "title": "Response List Sessions Sessions Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "sessions"
        ],
        "summary": "Create Session",
        "description": "Create a new session at L4 from a named demo dataset.",
        "operationId": "create_session_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/sessions/{session_id}/descend": {
      "post": {
        "tags": [
          "transitions"
        ],
        "summary": "Descend",
        "description": "Reduce complexity one level (L4\u2192L3\u2192\u2026\u2192L0).",
        "operationId": "descend_sessions__session_id__descend_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DescendRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sessions/{session_id}/ascend": {
      "post": {
        "tags": [
          "transitions"
        ],
        "summary": "Ascend",
        "description": "Increase complexity one level (L0\u2192L1\u2192L2\u2192L3; never L4).",
        "operationId": "ascend_sessions__session_id__ascend_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AscendRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AscendRequest": {
        "properties": {
          "enrichment_func": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enrichment Func",
            "description": "L0\u2192L1 enrichment name; omit to use parent-data 'source_expansion' / defaults."
          },
          "dimensions": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dimensions",
            "description": "L1\u2192L2 and L2\u2192L3 dimension names; omit for registry defaults."
          },
          "relationships": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relationships",
            "description": "L2\u2192L3 relationship names."
          },
          "source_column": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Column",
            "description": "L2\u2192L3 column to derive from (defaults to 'value')."
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "AscendRequest",
        "description": "`increase_complexity(target = current+1)`. Never reaches L4.\n\nField names match the engine's ascent adapter kwargs (`_ascent_params`);\nall are optional because each edge falls back to registry defaults (and L0\u2192L1\nreconstructs from the persisted parent vector \u2014 the spec-015 T007 fix)."
      },
...
      "CreateSessionRequest": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source",
            "description": "Dataset source, e.g. 'demo:school_scores' | 'demo:ademe' | 'demo:energy'.",
            "default": "demo:school_scores",
            "examples": [
              "demo:school_scores"
            ]
          }
        },
        "type": "object",
        "title": "CreateSessionRequest",
        "description": "Create a session from a named demo dataset (or, later, an upload)."
      },
...
      "DescendRequest": {
        "properties": {
          "builder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Builder",
            "description": "L4\u2192L3 graph builder name, e.g. 'rows_as_nodes' | 'bipartite'."
          },
          "config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config",
            "description": "Builder kwargs (e.g. {'id_column': 'id'})."
          },
          "domains": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domains",
            "description": "L3\u2192L2 category labels, e.g. ['high score','low score']."
          },
          "use_semantic": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Use Semantic",
            "description": "L3\u2192L2 use embedding similarity for matching."
          },
          "threshold": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Threshold",
            "description": "L3\u2192L2 semantic similarity threshold (0\u20131)."
          },
          "column": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Column",
            "description": "L2\u2192L1 column to keep."
          },
          "filter_query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filter Query",
            "description": "L2\u2192L1 optional pandas query filter."
          },
          "aggregation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aggregation",
            "description": "L1\u2192L0 aggregation: 'mean' | 'sum' | 'count' | ..."
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "DescendRequest",
        "description": "`reduce_complexity(target = current-1)`. Fields are per-edge; only the\nrelevant ones apply at each level (extras pass through to the engine)."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SessionState": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "current_level": {
            "type": "integer",
            "title": "Current Level",
            "description": "0\u20134; the level the session now sits at."
          },
          "current_node_id": {
            "type": "string",
            "title": "Current Node Id"
          },
          "summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Summary",
            "description": "Shape/counts of the current artifact; includes `value` at L0."
          },
          "available_moves": {
            "additionalProperties": {
              "items": {},
              "type": "array"
            },
            "type": "object",
            "title": "Available Moves",
            "description": "{'descend': [...], 'ascend': [...]} valid next transitions."
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "current_level",
          "current_node_id"
        ],
        "title": "SessionState",
        "description": "The projection every mutating/read call returns (see `state_of`)."
      },
      "SessionSummary": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "title"
        ],
        "title": "SessionSummary"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}